speech-api/historical.html

<!doctype html>
<title>Historical Speech API features</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
[
  "SpeechRecognitionError",
  "webkitSpeechGrammar",
  "webkitSpeechGrammarList",
  "webkitSpeechRecognition",
  "webkitSpeechRecognitionError",
  "webkitSpeechRecognitionEvent",
].forEach(name => {
  test(function() {
    assert_false(name in window);
  }, name + " interface should not exist");
});

// https://github.com/w3c/speech-api/pull/38
test(() => {
  assert_false("onmark" in SpeechSynthesisUtterance.prototype, "on prototype");
  assert_false("onmark" in new SpeechSynthesisUtterance, "on instance");
}, "onmark event handler attribute (for SSML) should not exist");
</script>
